Module monaco::sys

Bindings for the monaco namespace.

Modules

editor

Bindings for the monaco.editor namespace.

languages

Bindings for the monaco.languages namespace.

Structs

CancellationToken

CancellationTokenSource

Emitter

A helper that allows to emit and listen to typed events

Environment

IDisposable

IKeyboardEvent

IMarkdownString

IMouseEvent

IPosition

A position in the editor. This interface is suitable for serialization.

IRange

A range in the editor. This interface is suitable for serialization.

IScrollEvent

ISelection

A selection in the editor. The selection is a range that has an orientation.

KeyMod

Position

A position in the editor.

Range

A range in the editor. (startLineNumber,startColumn) is <= (endLineNumber,endColumn)

Selection

A selection in the editor. The selection is a range that has an orientation.

Token

Uri

Uniform Resource Identifier (Uri) http://tools.ietf.org/html/rfc3986. This class is a simple parser which creates the basic component parts (http://tools.ietf.org/html/rfc3986#section-3) with minimal validation and encoding.

UriComponents

Enums

KeyCode

MarkerSeverity

MarkerTag

SelectionDirection

Type Definitions

GetWorkerFn

Signature of the closure used by Environment::get_worker

GetWorkerUrlFn

Signature of the closure used by Environment::get_worker_url