Function monaco::sys::editor::create_web_worker

pub fn create_web_worker(opts: &IWebWorkerOptions) -> MonacoWebWorker

Create a new web worker that has model syncing capabilities built in. Specify an AMD module to load that will create an object that will be proxied.