Struct monaco::sys::editor::ISuggestOptions

#[repr(transparent)]
pub struct ISuggestOptions { /* private fields */ }

Configuration options for editor suggest widget

Implementations

impl ISuggestOptions

pub fn insert_mode(&self) -> Option<ISuggestOptionsInsertMode>

Overwrite word ends on accept. Default to false.

impl ISuggestOptions

pub fn set_insert_mode(&self, val: Option<ISuggestOptionsInsertMode>)

Set the insertMode property.

impl ISuggestOptions

pub fn insert_highlight(&self) -> Option<bool>

Show a highlight when suggestion replaces or keep text after the cursor. Defaults to false.

impl ISuggestOptions

pub fn set_insert_highlight(&self, val: Option<bool>)

Set the insertHighlight property.

impl ISuggestOptions

pub fn filter_graceful(&self) -> Option<bool>

Enable graceful matching. Defaults to true.

impl ISuggestOptions

pub fn set_filter_graceful(&self, val: Option<bool>)

Set the filterGraceful property.

impl ISuggestOptions

pub fn snippets_prevent_quick_suggestions(&self) -> Option<bool>

Prevent quick suggestions when a snippet is active. Defaults to true.

impl ISuggestOptions

pub fn set_snippets_prevent_quick_suggestions(&self, val: Option<bool>)

Set the snippetsPreventQuickSuggestions property.

impl ISuggestOptions

pub fn locality_bonus(&self) -> Option<bool>

Favours words that appear close to the cursor.

impl ISuggestOptions

pub fn set_locality_bonus(&self, val: Option<bool>)

Set the localityBonus property.

impl ISuggestOptions

pub fn share_suggest_selections(&self) -> Option<bool>

Enable using global storage for remembering suggestions.

impl ISuggestOptions

pub fn set_share_suggest_selections(&self, val: Option<bool>)

Set the shareSuggestSelections property.

impl ISuggestOptions

pub fn show_icons(&self) -> Option<bool>

Enable or disable icons in suggestions. Defaults to true.

impl ISuggestOptions

pub fn set_show_icons(&self, val: Option<bool>)

Set the showIcons property.

impl ISuggestOptions

pub fn max_visible_suggestions(&self) -> Option<f64>

Max suggestions to show in suggestions. Defaults to 12.

impl ISuggestOptions

pub fn set_max_visible_suggestions(&self, val: Option<f64>)

Set the maxVisibleSuggestions property.

impl ISuggestOptions

pub fn show_methods(&self) -> Option<bool>

Show method-suggestions.

impl ISuggestOptions

pub fn set_show_methods(&self, val: Option<bool>)

Set the showMethods property.

impl ISuggestOptions

pub fn show_functions(&self) -> Option<bool>

Show function-suggestions.

impl ISuggestOptions

pub fn set_show_functions(&self, val: Option<bool>)

Set the showFunctions property.

impl ISuggestOptions

pub fn show_constructors(&self) -> Option<bool>

Show constructor-suggestions.

impl ISuggestOptions

pub fn set_show_constructors(&self, val: Option<bool>)

Set the showConstructors property.

impl ISuggestOptions

pub fn show_fields(&self) -> Option<bool>

Show field-suggestions.

impl ISuggestOptions

pub fn set_show_fields(&self, val: Option<bool>)

Set the showFields property.

impl ISuggestOptions

pub fn show_variables(&self) -> Option<bool>

Show variable-suggestions.

impl ISuggestOptions

pub fn set_show_variables(&self, val: Option<bool>)

Set the showVariables property.

impl ISuggestOptions

pub fn show_classes(&self) -> Option<bool>

Show class-suggestions.

impl ISuggestOptions

pub fn set_show_classes(&self, val: Option<bool>)

Set the showClasses property.

impl ISuggestOptions

pub fn show_structs(&self) -> Option<bool>

Show struct-suggestions.

impl ISuggestOptions

pub fn set_show_structs(&self, val: Option<bool>)

Set the showStructs property.

impl ISuggestOptions

pub fn show_interfaces(&self) -> Option<bool>

Show interface-suggestions.

impl ISuggestOptions

pub fn set_show_interfaces(&self, val: Option<bool>)

Set the showInterfaces property.

impl ISuggestOptions

pub fn show_modules(&self) -> Option<bool>

Show module-suggestions.

impl ISuggestOptions

pub fn set_show_modules(&self, val: Option<bool>)

Set the showModules property.

impl ISuggestOptions

pub fn show_properties(&self) -> Option<bool>

Show property-suggestions.

impl ISuggestOptions

pub fn set_show_properties(&self, val: Option<bool>)

Set the showProperties property.

impl ISuggestOptions

pub fn show_events(&self) -> Option<bool>

Show event-suggestions.

impl ISuggestOptions

pub fn set_show_events(&self, val: Option<bool>)

Set the showEvents property.

impl ISuggestOptions

pub fn show_operators(&self) -> Option<bool>

Show operator-suggestions.

impl ISuggestOptions

pub fn set_show_operators(&self, val: Option<bool>)

Set the showOperators property.

impl ISuggestOptions

pub fn show_units(&self) -> Option<bool>

Show unit-suggestions.

impl ISuggestOptions

pub fn set_show_units(&self, val: Option<bool>)

Set the showUnits property.

impl ISuggestOptions

pub fn show_values(&self) -> Option<bool>

Show value-suggestions.

impl ISuggestOptions

pub fn set_show_values(&self, val: Option<bool>)

Set the showValues property.

impl ISuggestOptions

pub fn show_constants(&self) -> Option<bool>

Show constant-suggestions.

impl ISuggestOptions

pub fn set_show_constants(&self, val: Option<bool>)

Set the showConstants property.

impl ISuggestOptions

pub fn show_enums(&self) -> Option<bool>

Show enum-suggestions.

impl ISuggestOptions

pub fn set_show_enums(&self, val: Option<bool>)

Set the showEnums property.

impl ISuggestOptions

pub fn show_enum_members(&self) -> Option<bool>

Show enumMember-suggestions.

impl ISuggestOptions

pub fn set_show_enum_members(&self, val: Option<bool>)

Set the showEnumMembers property.

impl ISuggestOptions

pub fn show_keywords(&self) -> Option<bool>

Show keyword-suggestions.

impl ISuggestOptions

pub fn set_show_keywords(&self, val: Option<bool>)

Set the showKeywords property.

impl ISuggestOptions

pub fn show_words(&self) -> Option<bool>

Show text-suggestions.

impl ISuggestOptions

pub fn set_show_words(&self, val: Option<bool>)

Set the showWords property.

impl ISuggestOptions

pub fn show_colors(&self) -> Option<bool>

Show color-suggestions.

impl ISuggestOptions

pub fn set_show_colors(&self, val: Option<bool>)

Set the showColors property.

impl ISuggestOptions

pub fn show_files(&self) -> Option<bool>

Show file-suggestions.

impl ISuggestOptions

pub fn set_show_files(&self, val: Option<bool>)

Set the showFiles property.

impl ISuggestOptions

pub fn show_references(&self) -> Option<bool>

Show reference-suggestions.

impl ISuggestOptions

pub fn set_show_references(&self, val: Option<bool>)

Set the showReferences property.

impl ISuggestOptions

pub fn show_folders(&self) -> Option<bool>

Show folder-suggestions.

impl ISuggestOptions

pub fn set_show_folders(&self, val: Option<bool>)

Set the showFolders property.

impl ISuggestOptions

pub fn show_type_parameters(&self) -> Option<bool>

Show typeParameter-suggestions.

impl ISuggestOptions

pub fn set_show_type_parameters(&self, val: Option<bool>)

Set the showTypeParameters property.

impl ISuggestOptions

pub fn show_snippets(&self) -> Option<bool>

Show snippet-suggestions.

impl ISuggestOptions

pub fn set_show_snippets(&self, val: Option<bool>)

Set the showSnippets property.

impl ISuggestOptions

pub fn hide_status_bar(&self) -> Option<bool>

Controls the visibility of the status bar at the bottom of the suggest widget.

impl ISuggestOptions

pub fn set_hide_status_bar(&self, val: Option<bool>)

Set the hideStatusBar property.