Struct monaco::sys::editor::ITextModel

#[repr(transparent)]
pub struct ITextModel { /* private fields */ }

A model.

Implementations

impl ITextModel

pub fn uri(&self) -> Uri

Gets the resource associated with this editor model.

impl ITextModel

pub fn id(&self) -> String

A unique identifier associated with this model.

impl ITextModel

pub fn get_options(&self) -> TextModelResolvedOptions

Get the resolved options for this model.

impl ITextModel

pub fn get_version_id(&self) -> f64

Get the current version id of the model. Anytime a change happens to the model (even undo/redo), the version id is incremented.

impl ITextModel

pub fn get_alternative_version_id(&self) -> f64

Get the alternative version id of the model. This alternative version id is not always incremented, it will return the same values in the case of undo-redo.

impl ITextModel

pub fn set_value(&self, new_value: &str)

Replace the entire text buffer value contained in this model.

impl ITextModel

pub fn get_value(
&self,
eol: Option<EndOfLinePreference>,
preserve_bom: Option<bool>
) -> String

Get the text stored in this model. @param eol The end of line character preference. Defaults to EndOfLinePreference.TextDefined. @param preserverBOM Preserve a BOM character if it was detected when the model was constructed. @return The text.

impl ITextModel

pub fn get_value_length(
&self,
eol: Option<EndOfLinePreference>,
preserve_bom: Option<bool>
) -> f64

Get the length of the text stored in this model.

impl ITextModel

pub fn get_value_in_range(
&self,
range: &IRange,
eol: Option<EndOfLinePreference>
) -> String

Get the text in a certain range. @param range The range describing what text to get. @param eol The end of line character preference. This will only be used for multiline ranges. Defaults to EndOfLinePreference.TextDefined. @return The text.

impl ITextModel

pub fn get_value_length_in_range(&self, range: &IRange) -> f64

Get the length of text in a certain range. @param range The range describing what text length to get. @return The text length.

impl ITextModel

pub fn get_character_count_in_range(&self, range: &IRange) -> f64

Get the character count of text in a certain range. @param range The range describing what text length to get.

impl ITextModel

pub fn get_line_count(&self) -> f64

Get the number of lines in the model.

impl ITextModel

pub fn get_line_content(&self, line_number: f64) -> String

Get the text for a certain line.

impl ITextModel

pub fn get_line_length(&self, line_number: f64) -> f64

Get the text length for a certain line.

impl ITextModel

pub fn get_lines_content(&self) -> Array

Get the text for all lines.

Returns

string[]

impl ITextModel

pub fn get_eol(&self) -> String

Get the end of line sequence predominantly used in the text buffer. @return EOL char sequence (e.g.: ‘\n’ or ‘\r\n’).

impl ITextModel

pub fn get_line_min_column(&self, line_number: f64) -> f64

Get the minimum legal column for line at lineNumber

impl ITextModel

pub fn get_line_max_column(&self, line_number: f64) -> f64

Get the maximum legal column for line at lineNumber

impl ITextModel

pub fn get_line_first_non_whitespace_column(&self, line_number: f64) -> f64

Returns the column before the first non whitespace character for line at lineNumber. Returns 0 if line is empty or contains only whitespace.

impl ITextModel

pub fn get_line_last_non_whitespace_column(&self, line_number: f64) -> f64

Returns the column after the last non whitespace character for line at lineNumber. Returns 0 if line is empty or contains only whitespace.

impl ITextModel

pub fn validate_position(&self, position: &IPosition) -> Position

Create a valid position,

impl ITextModel

pub fn modify_position(
&self,
position: &IPosition,
offset: f64
) -> Result<Position, JsValue>

Advances the given position by the given offset (negative offsets are also accepted) and returns it as a new valid position.

If the offset and position are such that their combination goes beyond the beginning or end of the model, throws an exception.

If the offset is such that the new position would be in the middle of a multi-byte line terminator, throws an exception.

impl ITextModel

pub fn validate_range(&self, range: &IRange) -> Range

Create a valid range.

impl ITextModel

pub fn get_offset_at(&self, position: &IPosition) -> f64

Converts the position to a zero-based offset.

The position will be adjusted.

@param position A position. @return A valid zero-based offset.

impl ITextModel

pub fn get_position_at(&self, offset: f64) -> Position

Converts a zero-based offset to a position.

@param offset A zero-based offset. @return A valid position.

impl ITextModel

pub fn get_full_model_range(&self) -> Range

Get a range covering the entire model

impl ITextModel

pub fn is_disposed(&self) -> bool

Returns if the model was disposed or not.

impl ITextModel

pub fn find_matches(
&self,
search_string: &str,
search_only_editable_range: bool,
is_regex: bool,
match_case: bool,
word_separators: Option<&str>,
capture_matches: bool,
limit_result_count: Option<f64>
) -> Array

Search the model. @param searchString The string used to search. If it is a regular expression, set isRegex to true. @param searchOnlyEditableRange Limit the searching to only search inside the editable range of the model. @param isRegex Used to indicate that searchString is a regular expression. @param matchCase Force the matching to match lower/upper case exactly. @param wordSeparators Force the matching to match entire words only. Pass null otherwise. @param captureMatches The result will contain the captured groups. @param limitResultCount Limit the number of results @return The ranges where the matches are. It is empty if not matches have been found.

Returns

FindMatch[]

impl ITextModel

pub fn find_matches_range(
&self,
search_string: &str,
search_scope: &IRange,
is_regex: bool,
match_case: bool,
word_separators: Option<&str>,
capture_matches: bool,
limit_result_count: Option<f64>
) -> Array

Search the model. @param searchString The string used to search. If it is a regular expression, set isRegex to true. @param searchScope Limit the searching to only search inside this range. @param isRegex Used to indicate that searchString is a regular expression. @param matchCase Force the matching to match lower/upper case exactly. @param wordSeparators Force the matching to match entire words only. Pass null otherwise. @param captureMatches The result will contain the captured groups. @param limitResultCount Limit the number of results @return The ranges where the matches are. It is empty if no matches have been found.

Returns

FindMatch[]

impl ITextModel

pub fn find_next_match(
&self,
search_string: &str,
search_start: &IPosition,
is_regex: bool,
match_case: bool,
word_separators: Option<&str>,
capture_matches: bool
) -> Option<FindMatch>

Search the model for the next match. Loops to the beginning of the model if needed. @param searchString The string used to search. If it is a regular expression, set isRegex to true. @param searchStart Start the searching at the specified position. @param isRegex Used to indicate that searchString is a regular expression. @param matchCase Force the matching to match lower/upper case exactly. @param wordSeparators Force the matching to match entire words only. Pass null otherwise. @param captureMatches The result will contain the captured groups. @return The range where the next match is. It is null if no next match has been found.

impl ITextModel

pub fn find_previous_match(
&self,
search_string: &str,
search_start: &IPosition,
is_regex: bool,
match_case: bool,
word_separators: Option<&str>,
capture_matches: bool
) -> Option<FindMatch>

Search the model for the previous match. Loops to the end of the model if needed. @param searchString The string used to search. If it is a regular expression, set isRegex to true. @param searchStart Start the searching at the specified position. @param isRegex Used to indicate that searchString is a regular expression. @param matchCase Force the matching to match lower/upper case exactly. @param wordSeparators Force the matching to match entire words only. Pass null otherwise. @param captureMatches The result will contain the captured groups. @return The range where the previous match is. It is null if no previous match has been found.

impl ITextModel

pub fn get_mode_id(&self) -> String

Get the language associated with this model.

impl ITextModel

pub fn get_word_at_position(
&self,
position: &IPosition
) -> Option<IWordAtPosition>

Get the word under or besides position. @param position The position to look for a word. @return The word under or besides position. Might be null.

impl ITextModel

pub fn get_word_until_position(&self, position: &IPosition) -> IWordAtPosition

Get the word under or besides position trimmed to position.column @param position The position to look for a word. @return The word under or besides position. Will never be null.

impl ITextModel

pub fn delta_decorations(
&self,
old_decorations: &Array,
new_decorations: &Array,
owner_id: Option<f64>
) -> Array

Perform a minimum amount of operations, in order to transform the decorations identified by oldDecorations to the decorations described by newDecorations and returns the new identifiers associated with the resulting decorations.

@param oldDecorations Array containing previous decorations identifiers. @param newDecorations Array describing what decorations should result after the call. @param ownerId Identifies the editor id in which these decorations should appear. If no ownerId is provided, the decorations will appear in all editors that attach this model. @return An array containing the new decorations identifiers.

Arguments

old_decorations - string[]

new_decorations - IModelDeltaDecoration[]

Returns

string[]

impl ITextModel

pub fn get_decoration_options(&self, id: &str) -> Option<IModelDecorationOptions>

Get the options associated with a decoration. @param id The decoration id. @return The decoration options or null if the decoration was not found.

impl ITextModel

pub fn get_decoration_range(&self, id: &str) -> Option<Range>

Get the range associated with a decoration. @param id The decoration id. @return The decoration range or null if the decoration was not found.

impl ITextModel

pub fn get_line_decorations(
&self,
line_number: f64,
owner_id: Option<f64>,
filter_out_validation: Option<bool>
) -> Array

Gets all the decorations for the line lineNumber as an array. @param lineNumber The line number @param ownerId If set, it will ignore decorations belonging to other owners. @param filterOutValidation If set, it will ignore decorations specific to validation (i.e. warnings, errors). @return An array with the decorations

Returns

IModelDecoration[]

impl ITextModel

pub fn get_lines_decorations(
&self,
start_line_number: f64,
end_line_number: f64,
owner_id: Option<f64>,
filter_out_validation: Option<bool>
) -> Array

Gets all the decorations for the lines between startLineNumber and endLineNumber as an array. @param startLineNumber The start line number @param endLineNumber The end line number @param ownerId If set, it will ignore decorations belonging to other owners. @param filterOutValidation If set, it will ignore decorations specific to validation (i.e. warnings, errors). @return An array with the decorations

Returns

IModelDecoration[]

impl ITextModel

pub fn get_decorations_in_range(
&self,
range: &IRange,
owner_id: Option<f64>,
filter_out_validation: Option<bool>
) -> Array

Gets all the decorations in a range as an array. Only startLineNumber and endLineNumber from range are used for filtering. So for now it returns all the decorations on the same line as range. @param range The range to search in @param ownerId If set, it will ignore decorations belonging to other owners. @param filterOutValidation If set, it will ignore decorations specific to validation (i.e. warnings, errors). @return An array with the decorations

Returns

IModelDecoration[]

impl ITextModel

pub fn get_all_decorations(
&self,
owner_id: Option<f64>,
filter_out_validation: Option<bool>
) -> Array

Gets all the decorations as an array. @param ownerId If set, it will ignore decorations belonging to other owners. @param filterOutValidation If set, it will ignore decorations specific to validation (i.e. warnings, errors).

Returns

IModelDecoration[]

impl ITextModel

pub fn get_overview_ruler_decorations(
&self,
owner_id: Option<f64>,
filter_out_validation: Option<bool>
) -> Array

Gets all the decorations that should be rendered in the overview ruler as an array. @param ownerId If set, it will ignore decorations belonging to other owners. @param filterOutValidation If set, it will ignore decorations specific to validation (i.e. warnings, errors).

Returns

IModelDecoration[]

impl ITextModel

pub fn normalize_indentation(&self, str: &str) -> String

Normalize a string containing whitespace according to indentation rules (converts to spaces or to tabs).

impl ITextModel

pub fn update_options(&self, new_opts: &ITextModelUpdateOptions)

Change the options of this model.

impl ITextModel

pub fn detect_indentation(
&self,
default_insert_spaces: bool,
default_tab_size: f64
)

Detect the indentation options for this model from its content.

impl ITextModel

pub fn push_stack_element(&self)

Push a stack element onto the undo stack. This acts as an undo/redo point. The idea is to use pushEditOperations to edit the model and then to pushStackElement to create an undo/redo stop point.

impl ITextModel

pub fn push_edit_operations(
&self,
before_cursor_state: &Array,
edit_operations: &Array,
cursor_state_computer: Option<&Array>
) -> Option<Array>

Push edit operations, basically editing the model. This is the preferred way of editing the model. The edit operations will land on the undo stack. @param beforeCursorState The cursor state before the edit operations. This cursor state will be returned when undo or redo are invoked. @param editOperations The edit operations. @param cursorStateComputer A callback that can compute the resulting cursors state after the edit operations have been executed. @return The cursor state returned by the cursorStateComputer.

Arguments

before_cursor_state - Selection[]

edit_operations - IIdentifiedSingleEditOperation[]

cursor_state_computer - ( inverseEditOperations: IIdentifiedSingleEditOperation[] ) => Selection[]

Returns

Selection[]

impl ITextModel

pub fn push_eol(&self, eol: EndOfLineSequence)

Change the end of line sequence. This is the preferred way of changing the eol sequence. This will land on the undo stack.

impl ITextModel

pub fn apply_edits(&self, operations: &Array) -> Array

Edit the model without adding the edits to the undo stack. This can have dire consequences on the undo stack! See @pushEditOperations for the preferred way. @param operations The edit operations. @return The inverse edit operations, that, when applied, will bring the model back to the previous state.

Arguments

operations - IIdentifiedSingleEditOperation[]

Returns

IIdentifiedSingleEditOperation[]

impl ITextModel

pub fn set_eol(&self, eol: EndOfLineSequence)

Change the end of line sequence without recording in the undo stack. This can have dire consequences on the undo stack! See @pushEOL for the preferred way.

impl ITextModel

pub fn on_did_change_content(&self, listener: &Function) -> IDisposable

An event emitted when the contents of the model have changed. @event

Arguments

listener - (e: IModelContentChangedEvent) => void

impl ITextModel

pub fn on_did_change_decorations(&self, listener: &Function) -> IDisposable

An event emitted when decorations of the model have changed. @event

Arguments

listener - (e: {}) => void

impl ITextModel

pub fn on_did_change_options(&self, listener: &Function) -> IDisposable

An event emitted when the model options have changed. @event

Arguments

listener - (e: IModelOptionsChangedEvent) => void

impl ITextModel

pub fn on_did_change_language(&self, listener: &Function) -> IDisposable

An event emitted when the language associated with the model has changed. @event

Arguments

listener - (e: IModelLanguageChangedEvent) => void

impl ITextModel

pub fn on_did_change_language_configuration(
&self,
listener: &Function
) -> IDisposable

An event emitted when the language configuration associated with the model has changed. @event

Arguments

listener - (e: {}) => void

impl ITextModel

pub fn on_will_dispose(&self, listener: &Function) -> IDisposable

An event emitted right before disposing the model. @event

Arguments

listener - () => void

impl ITextModel

pub fn dispose(&self)

Destroy this model. This will unbind the model from the mode and make all necessary clean-up to release this object to the GC.