Enum monaco::sys::KeyCode

pub enum KeyCode {

    Unknown,
    Backspace,
    Tab,
    Enter,
    Shift,
    Ctrl,
    Alt,
    Pausebreak,
    Capslock,
    Escape,
    Space,
    Pageup,
    Pagedown,
    End,
    Home,
    Leftarrow,
    Uparrow,
    Rightarrow,
    Downarrow,
    Insert,
    Delete,
    Key0,
    Key1,
    Key2,
    Key3,
    Key4,
    Key5,
    Key6,
    Key7,
    Key8,
    Key9,
    KeyA,
    KeyB,
    KeyC,
    KeyD,
    KeyE,
    KeyF,
    KeyG,
    KeyH,
    KeyI,
    KeyJ,
    KeyK,
    KeyL,
    KeyM,
    KeyN,
    KeyO,
    KeyP,
    KeyQ,
    KeyR,
    KeyS,
    KeyT,
    KeyU,
    KeyV,
    KeyW,
    KeyX,
    KeyY,
    KeyZ,
    Meta,
    Contextmenu,
    F1,
    F2,
    F3,
    F4,
    F5,
    F6,
    F7,
    F8,
    F9,
    F10,
    F11,
    F12,
    F13,
    F14,
    F15,
    F16,
    F17,
    F18,
    F19,
    Numlock,
    Scrolllock,
    UsSemicolon,
    UsEqual,
    UsComma,
    UsMinus,
    UsDot,
    UsSlash,
    UsBacktick,
    UsOpenSquareBracket,
    UsBackslash,
    UsCloseSquareBracket,
    UsQuote,
    Oem8,
    Oem102,
    Numpad0,
    Numpad1,
    Numpad2,
    Numpad3,
    Numpad4,
    Numpad5,
    Numpad6,
    Numpad7,
    Numpad8,
    Numpad9,
    NumpadMultiply,
    NumpadAdd,
    NumpadSeparator,
    NumpadSubtract,
    NumpadDecimal,
    NumpadDivide,
    KeyInComposition,
    AbntC1,
    AbntC2,
    MaxValue,

}

Variants

Unknown

Placed first to cover the 0 value of the enum.

Backspace

Tab

Enter

Shift

Ctrl

Alt

Pausebreak

Capslock

Escape

Space

Pageup

Pagedown

End

Home

Leftarrow

Uparrow

Rightarrow

Downarrow

Insert

Delete

Key0

Key1

Key2

Key3

Key4

Key5

Key6

Key7

Key8

Key9

KeyA

KeyB

KeyC

KeyD

KeyE

KeyF

KeyG

KeyH

KeyI

KeyJ

KeyK

KeyL

KeyM

KeyN

KeyO

KeyP

KeyQ

KeyR

KeyS

KeyT

KeyU

KeyV

KeyW

KeyX

KeyY

KeyZ

Meta

Contextmenu

F1

F2

F3

F4

F5

F6

F7

F8

F9

F10

F11

F12

F13

F14

F15

F16

F17

F18

F19

Numlock

Scrolllock

UsSemicolon

Used for miscellaneous characters; it can vary by keyboard. For the US standard keyboard, the ‘;:’ key

UsEqual

For any country/region, the ‘+’ key For the US standard keyboard, the ‘=+’ key

UsComma

For any country/region, the ‘,’ key For the US standard keyboard, the ‘,<’ key

UsMinus

For any country/region, the ‘-’ key For the US standard keyboard, the ‘-_’ key

UsDot

For any country/region, the ‘.’ key For the US standard keyboard, the ‘.>’ key

UsSlash

Used for miscellaneous characters; it can vary by keyboard. For the US standard keyboard, the ‘/?’ key

UsBacktick

Used for miscellaneous characters; it can vary by keyboard. For the US standard keyboard, the ‘`~’ key

UsOpenSquareBracket

Used for miscellaneous characters; it can vary by keyboard. For the US standard keyboard, the ‘[{’ key

UsBackslash

Used for miscellaneous characters; it can vary by keyboard. For the US standard keyboard, the ‘|’ key

UsCloseSquareBracket

Used for miscellaneous characters; it can vary by keyboard. For the US standard keyboard, the ‘]}’ key

UsQuote

Used for miscellaneous characters; it can vary by keyboard. For the US standard keyboard, the ‘’“’ key

Oem8

Used for miscellaneous characters; it can vary by keyboard.

Oem102

Either the angle bracket key or the backslash key on the RT 102-key keyboard.

Numpad0

Numpad1

Numpad2

Numpad3

Numpad4

Numpad5

Numpad6

Numpad7

Numpad8

Numpad9

NumpadMultiply

NumpadAdd

NumpadSeparator

NumpadSubtract

NumpadDecimal

NumpadDivide

KeyInComposition

Cover all key codes when IME is processing input.

AbntC1

AbntC2

MaxValue

Placed last to cover the length of the enum. Please do not depend on this value!

Implementations

impl KeyCode

pub fn from_value(val: u32) -> Option<Self>

Get the variant for the value. Returns None if the value isn’t part of the enum.

pub fn to_value(&self) -> u32

Get the value of the variant.