Struct monaco::sys::languages::DeclarationProvider

#[repr(transparent)]
pub struct DeclarationProvider { /* private fields */ }

The definition provider interface defines the contract between extensions and the go to definition and peek definition features.

Implementations

impl DeclarationProvider

pub fn provide_declaration(
&self,
model: &ITextModel,
position: &Position,
token: &CancellationToken
) -> JsValue

Provide the declaration of the symbol at the given position and document.