Module monaco::api

Simplified API. Requires the “api” feature (enabled by default).

Most of the types here are simply wrappers around the Javascript types and implement AsRef for them.

Structs

CodeEditor

Monaco code editor.

CodeEditorOptions

Options for creating a new editor. This represents a simplified version of IStandaloneEditorConstructionOptions.

DisposableClosure

A Closure that is tied to an IDisposable.

TextModel

Models are a more abstract representation of files that can be “opened” (attached) to an editor.

Functions

set_global_builtin_theme

Switches to a built-in theme.

set_global_theme

Switches to a theme.