Module monaco::sys::editor

Bindings for the monaco.editor namespace.

Structs

BareFontInfo

ConfigurationChangedEvent

An event describing that the configuration of the editor has changed.

EditorLayoutInfo

The internal layout details of the editor.

EditorWrappingInfo

FindMatch

FontInfo

IActionDescriptor

Description of an action contribution

IChange

A change

ICharChange

A character level change.

ICodeEditor

A rich code editor.

ICodeEditorViewState

A (serializable) state of the code editor.

IColorizerElementOptions

IColorizerOptions

ICommand

A command that modifies text / cursor state on a model.

IComputedEditorOptions

All computed editor options.

IContentSizeChangedEvent

IContentWidget

A content widget renders inline with the text and can be easily placed ‘near’ an editor position.

IContentWidgetPosition

A position for rendering content widgets.

IContextKey

ICursorPositionChangedEvent

An event describing that the cursor position has changed.

ICursorSelectionChangedEvent

An event describing that the cursor selection has changed.

ICursorState

A (serializable) state of the cursors.

ICursorStateComputerData

A helper for computing cursor state after a command.

IDecorationOptions

IDiffEditor

A rich diff editor.

IDiffEditorConstructionOptions

The options to create a diff editor.

IDiffEditorModel

A model for the diff editor.

IDiffEditorOptions

Configuration options for the diff editor.

IDiffEditorViewState

(Serializable) View state for the diff editor.

IDiffLineInformation

Information about a line in the diff editor

IDiffNavigator

IDiffNavigatorOptions

IDimension

IEditOperationBuilder

A builder and helper for edit operations for a command.

IEditor

An editor.

IEditorAction

IEditorCommentsOptions

Configuration options for editor comments

IEditorConstructionOptions

IEditorContribution

An editor contribution that gets created every time a new editor gets created and gets disposed when the editor gets disposed.

IEditorFindOptions

Configuration options for editor find widget

IEditorHoverOptions

Configuration options for editor hover

IEditorLightbulbOptions

Configuration options for editor lightbulb

IEditorMinimapOptions

Configuration options for editor minimap

IEditorMouseEvent

A mouse event originating from the editor.

IEditorOption

IEditorOptions

Configuration options for the editor.

IEditorParameterHintOptions

Configuration options for parameter hints

IEditorScrollbarOptions

Configuration options for editor scrollbars

IGlobalEditorOptions

Options which apply for all editors.

IGotoLocationOptions

Configuration options for go to location

IIdentifiedSingleEditOperation

A single edit operation, that has an identifier.

ILineChange

A line change

IMarker

IMarkerData

A structure defining a problem/warning/etc.

IModelChangedEvent

An event describing that an editor has had its model reset (i.e. editor.setModel()).

IModelContentChange

IModelContentChangedEvent

An event describing a change in the text of a model.

IModelDecoration

A decoration in the model.

IModelDecorationMinimapOptions

Options for rendering a model decoration in the overview ruler.

IModelDecorationOptions

Options for a model decoration.

IModelDecorationOverviewRulerOptions

Options for rendering a model decoration in the overview ruler.

IModelDeltaDecoration

New model decorations.

IModelLanguageChangedEvent

An event describing that the current mode associated with a model has changed.

IModelOptionsChangedEvent

IMouseTarget

Target hit with the mouse in the editor.

INewScrollPosition

IOverlayWidget

An overlay widgets renders on top of the text.

IOverlayWidgetPosition

A position for rendering overlay widgets.

IPartialEditorMouseEvent

IPasteEvent

A paste event originating from the editor.

IQuickSuggestionsOptions

Configuration options for quick suggestions

IRelatedInformation

ISingleEditOperation

A single edit operation, that acts as a simple replace. i.e. Replace text at range with text in model.

IStandaloneCodeEditor

IStandaloneDiffEditor

IStandaloneEditorConstructionOptions

The options to create an editor.

IStandaloneThemeData

ISuggestOptions

Configuration options for editor suggest widget

ITextModel

A model.

ITextModelUpdateOptions

ITokenThemeRule

IViewState

A (serializable) state of the view.

IViewZone

A view zone is a full horizontal rectangle that ‘pushes’ text down. The editor reserves space for view zones when rendering.

IViewZoneChangeAccessor

An accessor that allows for zones to be added or removed.

IWebWorkerOptions

IWordAtPosition

Word inside a model.

InternalEditorRenderLineNumbersOptions

InternalEditorScrollbarOptions

MonacoWebWorker

A web worker that can provide a proxy to an arbitrary file.

OverviewRulerPosition

A description for the overview ruler position.

TextModelResolvedOptions

ThemeColor

Enums

AccessibilitySupport

BuiltinTheme

ContentWidgetPositionPreference

CursorChangeReason

DefaultEndOfLine

EditorAutoClosingOvertypeStrategy

EditorAutoClosingStrategy

EditorAutoIndentStrategy

EditorAutoSurroundStrategy

EditorOption

EndOfLinePreference

EndOfLineSequence

GoToLocationValues

IEditorFindOptionsAutoFindInSelection

IEditorMinimapOptionsShowSlider

IEditorMinimapOptionsSide

IEditorOptionsAcceptSuggestionOnEnter

IEditorOptionsAccessibilitySupport

IEditorOptionsAutoIndent

IEditorOptionsCursorBlinking

IEditorOptionsCursorStyle

IEditorOptionsCursorSurroundingLinesStyle

IEditorOptionsFoldingStrategy

IEditorOptionsMatchBrackets

IEditorOptionsMouseStyle

IEditorOptionsMultiCursorModifier

IEditorOptionsMultiCursorPaste

IEditorOptionsPeekWidgetDefaultFocus

IEditorOptionsRenderLineHighlight

IEditorOptionsRenderValidationDecorations

IEditorOptionsRenderWhitespace

IEditorOptionsShowFoldingControls

IEditorOptionsSnippetSuggestions

IEditorOptionsSuggestSelection

IEditorOptionsTabCompletion

IEditorOptionsWordWrap

IEditorOptionsWrappingIndent

IEditorOptionsWrappingStrategy

IEditorScrollbarOptionsHorizontal

IEditorScrollbarOptionsVertical

ISuggestOptionsInsertMode

LineNumbersType

MinimapPosition

MouseTargetType

OverlayWidgetPositionPreference

OverviewRulerLane

RenderLineNumbersType

RenderMinimap

ScrollType

ScrollbarVisibility

TextEditorCursorBlinkingStyle

TextEditorCursorStyle

TrackedRangeStickiness

WrappingIndent

Functions

colorize

Colorize text using language languageId.

colorize_element

Colorize the contents of domNode using attribute data-lang.

colorize_model_line

Colorize a line in a model.

create

Create a new editor under domElement. domElement should be empty (not contain other dom nodes). The editor will read the size of domElement.

create_diff_editor

Create a new diff editor under domElement. domElement should be empty (not contain other dom nodes). The editor will read the size of domElement.

create_diff_navigator

create_model

Create a new editor model. You can specify the language that should be set for this model or let the language be inferred from the uri.

create_web_worker

Create a new web worker that has model syncing capabilities built in. Specify an AMD module to load that will create an object that will be proxied.

define_theme

Define a new theme or update an existing theme.

get_model

Get the model that has uri if it exists.

get_model_markers

Get markers for owner and/or resource

get_models

Get all the created models.

on_did_change_model_language

Emitted when a different language is set to a model. @event

on_did_create_editor

Emitted when an editor is created. Creating a diff editor might cause this listener to be invoked with the two editors. @event

on_did_create_model

Emitted when a model is created. @event

on_will_dispose_model

Emitted right before a model is disposed. @event

remeasure_fonts

Clears all cached font measurements and triggers re-measurement.

set_model_language

Change the language for a model.

set_model_markers

Set the markers for a model.

set_theme

Switches to a theme.

tokenize

Tokenize text using language languageId